+++ klause/src/commonMain/kotlin/com/eignex/klause/backtrack/selector/IndomainSet.kt
@@ -7,9 +7,9 @@ import kotlin.random.Random
  * Allow-list value selection: tries only [allowedValues] (in order) intersected with the
  * current domain. Sparse-aware via the `in d` membership check.
  */
-internal class IndomainSet(private val allowedValues: IntArray) : ValueSelector {
-    override fun values(session: PropagationSession, varRef: VarRef, rng: Random): Sequence<Int> = when (varRef) {
-        is VarRef.Bool -> allowedValues.asSequence().filter { it == 0 || it == 1 }
+internal class IndomainSet(private val allowedValues: LongArray) : ValueSelector {
+    override fun values(session: PropagationSession, varRef: VarRef, rng: Random): Sequence<Long> = when (varRef) {
+        is VarRef.Bool -> allowedValues.asSequence().filter { it == 0L || it == 1L }
 
         is VarRef.IntVar -> {
             val d = session.intDomain(varRef.varId)
